third_party/blink/web_tests/platform/generic/fast/url/trivial-expected.txt

Test basic features of URL canonicalization

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS canonicalize('http://example.com/') is 'http://example.com/'
PASS canonicalize('/') is 'http://example.org/'
PASS successfullyParsed is true

TEST COMPLETE
